### 2025-04-29 #8132 @editorialbot
[PRE REVIEW]: Dfitspy
**Submitting author:** @Romain-Thomas-Shef (<a href="http://orcid.org/0000-0001-8385-3276">Romain Thomas</a>)
**Repository:** https://github.com/Romain-Thomas-Shef/dfitspy
**Branch with paper.md** (empty if default branch):
**Version:** 25.1.1
