### 2025-04-29 astrom-tom/dfitspy#21 @EdwinB12
Add a contribution guide to the readme
Add some instructions for people to contribute to the package
