### 2025-04-29 #8132 @arfon
@edwinb12 would you be willing to review this submission for JOSS?
