### 2025-04-29 #8133 @EdwinB12
@Romain-Thomas-Shef - This is a great submission! I have some small suggestions for things you could change.

I would recommend adding a contributor guide to your readme. I have made an issue at https://github.com/Romain-Thomas-Shef/dfitspy/issues/21
